rise: fixed CEvent for pthread version < 0.10

+++ rise/rise/os/osthread.h
@@ -31,7 +31,7 @@
 #include <sys/time.h>
 #endif
 
-#define RISE_WAIT_INFINITE -1
+#define RISE_WAIT_INFINITE static_cast<unsigned long>(-1)
 
 namespace rise
 {
